Fix #621: Step In on a line with breakpoint breaks things
Introduce the concept of a "protected environment" - use a throw-away dummy environment as evaluation context, with designated target environment set as a parent of the dummy, to avoid issues when debug bit is set on the environment (e.g. after stepping). Evaluations are protected by default, but a flag to request unprotected eval is provided.